Add load_jQuery function
load_jQuery injects jQuery from the google CDN if it's not found, waits for jQuery to load, and then runs a callback function when jQuery is ready to go.

This function can then be used in conjunction with a setup function to automatically load jQuery before the script executes its main logic.